pyproject.toml
[project]
name = "langgraph-study"
version = "0.1.0"
description = "LangGraph study project"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=1.1.4",
    "langgraph>=1.0.5",
    "marimo>=0.18.4",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]
